added table variable to models for use $table variable added custom js for tiny mce init handler modify migrations and seeds override default bread views for customize added link to relations in formfields.relationship view

+++ database/migrations/2020_04_15_173313_create_page_to_tag_table.php
@@ -15,7 +15,7 @@ class CreatePageToTagTable extends Migration
     {
         Schema::create('page_to_tag', function (Blueprint $table) {
             $table->id();
-            $table->integer('page_id')->unsigned();
+            $table->bigInteger('page_id')->unsigned();
             $table->foreign('page_id')->references('id')->on('pages')->onDelete('cascade');
             $table->bigInteger('tag_id')->unsigned();
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
